cosmetics

+++ invoices/action.py
@@ -217,6 +217,9 @@ def cleanup_contacts(modeladmin, request, queryset):
         delete_all_contacts.delay(queryset)
 
 def cleanup_some_contacts(modeladmin, request, queryset):
+    if not request.user.is_superuser:
+        modeladmin.message_user(request, _("You are not allowed to do this action"))
+        return
     from invoices.processors.tasks import delete_some_contacts
     if os.environ.get('LOCAL_ENV', None):
         delete_some_contacts(queryset)
